+++ backend/src/Controller/Api/V1/SkillController.php
@@ -147,6 +147,15 @@ class SkillController extends AbstractController
      *      @OA\Property(property="message", type="string")
      *     )
      * )
+     * @OA\Response(
+     *     response=403,
+     *     description="Return a unAuthorization message",
+     *     @OA\Schema(
+     *      type="object",
+     *      @OA\Property(property="status", type="integer"),
+     *      @OA\Property(property="message", type="string")
+     *     )
+     * )
      * @OA\Parameter(
      *     name="skill",
      *     in="body",
@@ -162,6 +171,8 @@ class SkillController extends AbstractController
     public function edit(Request $request, Skill $skill, ValidatorInterface $validator)
     {
         try {
+            //! Voter
+            $this->denyAccessUnlessGranted('EDIT', $skill);
 
             $jsonData = json_decode($request->getContent());
     
@@ -218,10 +229,22 @@ class SkillController extends AbstractController
      *      @OA\Property(property="message", type="string")
      *     )
      * )
+     * @OA\Response(
+     *     response=403,
+     *     description="Return a unAuthorization message",
+     *     @OA\Schema(
+     *      type="object",
+     *      @OA\Property(property="status", type="integer"),
+     *      @OA\Property(property="message", type="string")
+     *     )
+     * )
      * @Route("/{id}", name="delete", methods={"DELETE"}, requirements={"id": "\d+"})
      */
     public function delete(Skill $skill)
     {
+        //! Voter
+        $this->denyAccessUnlessGranted('DELETE', $skill);
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($skill);
         $em->flush();
